oura-ai/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Martin Nolan",email = "martinnolan_1@live.co.uk"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "httpx[http2] (>=0.28.1,<0.29.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
black = "^25.1.0"
isort = "^6.0.1"
ruff = "^0.11.10"

